chainlib/eth/address.py
# third-party imports
import sha3
from hexathon import (
        strip_0x,
        uniform,
    )
from crypto_dev_signer.encoding import (
        is_address,
        is_checksum_address,
        to_checksum_address,
        )

to_checksum = to_checksum_address
