requirements.txt
crypto-dev-signer~=0.4.14b6
pysha3==1.0.2
hexathon~=0.0.1a7
websocket-client==0.57.0
potaahto~=0.0.1a1
